+++ src/main/kotlin/yqloss/yqlossclientmixinkt/module/rawinput/NativeRawInputProvider.kt
@@ -19,6 +19,7 @@
 package yqloss.yqlossclientmixinkt.module.rawinput
 
 import org.lwjgl.input.Mouse
+import org.lwjgl.opengl.Display
 import yqloss.yqlossclientmixinkt.nativeapi.registerRawInputDevices
 import yqloss.yqlossclientmixinkt.nativeapi.unregisterRawInputDevices
 import yqloss.yqlossclientmixinkt.util.property.versionedLazy
@@ -27,6 +28,10 @@ import yqloss.yqlossclientmixinkt.util.scope.nothrow
 object NativeRawInputProvider : RawInputProvider {
     private val onGrabStateChange: Unit by versionedLazy({ RawInput.provider === this && Mouse.isGrabbed() }) {
         nothrow {
+            val grabbed = Mouse.isGrabbed()
+            Mouse.setCursorPosition(Display.getWidth() / 2, Display.getHeight() / 2)
+            Mouse.setGrabbed(!grabbed)
+            Mouse.setGrabbed(grabbed)
             if (RawInput.provider === this && Mouse.isGrabbed()) {
                 registerRawInputDevices()
             } else {